move RotFlip, Tile and TileConflictRule to new file
